Updated favouriteBooks with correct code and deleted testBookSearch
